server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
ZODB==5.8.1
persistent==5.0
transaction==3.1.0
BTrees==4.11.3
pydantic==2.4.2
geopy==2.3.0
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0